refactoring new folder structure with color extractor
