+++ src/xrl/agents/env_steps.py
@@ -2,11 +2,14 @@
 # DELETE ME #
 #############
 
+# TODO: Finish refactoring, so that this file is no longer needed!!!!
+
+
 # main file which is selecting the raw features extractor 
 # and the processer to meaninful features
 
-import xrl.features.aari_raw_features_extractor as aari_raw_features_extractor
-import xrl.features.aari_feature_processer as aari_features_processer
+import xrl.agents.image_extractors.aari_raw_features_extractor as aari_raw_features_extractor
+import xrl.agents.feature_processing.aari_feature_processer as aari_features_processer
 
 
 # helper function to get features
@@ -25,5 +28,5 @@ def do_step(env, action=1, last_raw_features=None, raw_features_processer="atari
     # calculate meaningful features from given raw features and last raw features
     raw_features = aari_raw_features_extractor.get_raw_features(info, last_raw_features, gametype=gametype)
     raw_features, features = aari_features_processer.preprocess_raw_features(raw_features)
-    raise RuntimeError("Looks like refactoring is not complete for this call :( ...")
+    #raise RuntimeError("Looks like refactoring is not complete for this call :( ...")
     return raw_features, features, reward, done
